% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contributions.R
\name{calculate_contributions}
\alias{calculate_contributions}
\title{Function to calculate particulate matter relative contributions.}
\usage{
calculate_contributions(df)
}
\arguments{
\item{df}{Input data frame.}
}
\value{
Tibble.
}
\description{
Function to calculate particulate matter relative contributions.
}
\examples{

# Example input
data_example <- tibble::tribble(
 ~variable,        ~value,
 "mass",           39.7,  
 "ec",             6.1,   
 "om",             9.2,   
 "nitrate",        3.3,   
 "ammonium",       1.4,   
 "sulfate",        3.3,   
 "mineral_dust",   2.7,   
 "trace_elements", 6.8,
 "missing",        6.9
 )
 
# Determine missing mass
calculate_contributions(data_example)

}
\seealso{
\code{\link{calculate_missing_mass}}, 
\code{\link{str_contribution_label}}
}
\author{
Stuart K. Grange
}
