% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_missing_mass.R
\name{calculate_missing_mass}
\alias{calculate_missing_mass}
\title{Function to calculate particulate matter missing mass.}
\usage{
calculate_missing_mass(df, digits = 6, add_method_type = TRUE)
}
\arguments{
\item{df}{Input data frame.}

\item{digits}{Number of digits for logic test.}

\item{add_method_type}{Should a \code{method_type} variable be added to the 
return?}
}
\value{
Tibble.
}
\description{
\code{\link{calculate_missing_mass}} will add one more observation/row to
the input data frame/tibble.
}
\examples{

# Example input
data_example <- tibble::tribble(
 ~variable,        ~value,
 "mass",           39.7,  
 "elemental_carbon", 6.1,   
 "organic_matter", 9.2,   
 "nitrate",        3.3,   
 "ammonium",       1.4,   
 "sulfate",        3.3,   
 "mineral_dust",   2.7,   
 "trace_elements", 6.8
 )
 
# Determine missing mass
calculate_missing_mass(data_example)
calculate_missing_mass(data_example, add_method_type = FALSE)

}
\seealso{
\code{\link{calculate_contributions}}
}
\author{
Stuart K. Grange
}
