% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_measurements}
\alias{import_measurements}
\alias{import_field_campaigns}
\alias{import_sites}
\alias{import_mass_contributions}
\alias{import_data_sources}
\alias{import_laboratories}
\title{Functions to import data from a \strong{pmcharacterr} database.}
\usage{
import_measurements(
  con,
  organic_carbon_multiplier = 1,
  convert_units = FALSE,
  tz = "UTC"
)

import_field_campaigns(con, tz = "UTC")

import_sites(con)

import_mass_contributions(con)

import_data_sources(con)

import_laboratories(con)
}
\arguments{
\item{con}{Database connection to a \strong{pmcharacterr} database.}

\item{organic_carbon_multiplier}{Multiplier for \code{organic_carbon} 
measurements to be transformed to organic matter. Usually, a value of 
\code{1.6} or \code{1.8} is used. The variable name will also be switched to 
\code{organic_matter}.}

\item{convert_units}{Should units be normalised to \code{ug.m-3}. This 
requires other units to be set correctly in the \code{measurements} table.}

\item{tz}{Time zone to format dates to.}
}
\value{
Tibble.
}
\description{
Functions to import data from a \strong{pmcharacterr} database.
}
\author{
Stuart K. Grange
}
