% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_priority_variables.R
\name{select_priority_variables}
\alias{select_priority_variables}
\title{Function to select a collection of priority variables from a 
\code{`measurements`} \strong{pmcharacter} table.}
\usage{
select_priority_variables(df)
}
\arguments{
\item{df}{Tibble from \code{\link{import_measurements}}.}
}
\value{
Tibble.
}
\description{
Function to select a collection of priority variables from a 
\code{`measurements`} \strong{pmcharacter} table.
}
\author{
Stuart K. Grange
}
