% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contributions.R
\name{str_contribution_label}
\alias{str_contribution_label}
\title{Function to calculate particulate matter contribution labels.}
\usage{
str_contribution_label(x, digits = 1)
}
\arguments{
\item{x}{A numeric vector to format.}

\item{digits}{Number of decimal points to display in the 
\code{contribution_label} variable.}
}
\value{
Character vector.
}
\description{
Function to calculate particulate matter contribution labels.
}
\seealso{
\code{\link{calculate_missing_mass}}
}
\author{
Stuart K. Grange
}
