% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sulfur_to_sulfate.R
\name{sulfur_to_sulfate}
\alias{sulfur_to_sulfate}
\alias{nitrogen_to_nitrate}
\title{Functions to convert elements to ions.}
\usage{
sulfur_to_sulfate(x, s_mass = 32.06, o_mass = 15.999)

nitrogen_to_nitrate(x, n_mass = 14.007, o_mass = 15.999)
}
\arguments{
\item{x}{Numeric vector.}

\item{s_mass, o_mass, n_mass}{Atomic weight of elements.}
}
\value{
Numeric vector.
}
\description{
Functions to convert elements to ions.
}
\examples{

sulfur_to_sulfate(0.28)

}
\author{
Stuart K. Grange
}
