% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours_pmcharacterr.R
\name{colours_pmcharacterr}
\alias{colours_pmcharacterr}
\title{Function to generate a custom colour palette for particulate matter plots.}
\usage{
colours_pmcharacterr(as_palette = TRUE)
}
\arguments{
\item{as_palette}{Should the return be of class \code{palette}?}
}
\value{
Named colour palette or tibble.
}
\description{
Function to generate a custom colour palette for particulate matter plots.
}
\examples{

colours_pmcharacterr(as_palette = TRUE)
colours_pmcharacterr(as_palette = FALSE)

}
\author{
Stuart K. Grange
}
