% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_functions.R
\name{import_measurements}
\alias{import_measurements}
\alias{import_field_campaigns}
\alias{import_sites}
\alias{import_mass_contributions}
\alias{import_data_sources}
\alias{import_laboratories}
\alias{import_complete_dates}
\alias{import_error_summaries}
\title{Functions to import data from a \strong{pmcharacterr} database.}
\usage{
import_measurements(
  con,
  organic_carbon_multiplier = 1,
  convert_units = FALSE,
  value_type = NA,
  use_complete_dates = FALSE,
  tz = "UTC"
)

import_field_campaigns(con, tz = "UTC")

import_sites(con)

import_mass_contributions(con)

import_data_sources(con)

import_laboratories(con)

import_complete_dates(con, tz = "UTC")

import_error_summaries(con)
}
\arguments{
\item{con}{Database connection to a \strong{pmcharacterr} database.}

\item{organic_carbon_multiplier}{Multiplier for \code{organic_carbon} 
measurements to be transformed to organic matter. Usually, a value of 
\code{1.6} or \code{1.8} is used. The variable name will also be switched to 
\code{organic_matter}.}

\item{convert_units}{Should units be normalised to \code{ug.m-3}. This 
requires other units to be set correctly in the \code{measurements} table.}

\item{value_type}{A vector to filter the `value_type` variable to.}

\item{use_complete_dates}{If a `complete_observation_dates` table is present, 
should the observations be filtered to these dates?}

\item{tz}{Time zone to format dates to.}
}
\value{
Tibble.
}
\description{
Functions to import data from a \strong{pmcharacterr} database.
}
\author{
Stuart K. Grange
}
